GPII-1878: Updated query encoding and decoding to use static functions from `gpii-express`.
