+++ tests/js/lib/harness.js
@@ -1,13 +1,14 @@
+/*
+
+    Our test harness configuration.
+
+ */
 /* eslint-env node */
 "use strict";
 var fluid = require("infusion");
-var gpii  = fluid.registerNamespace("gpii");
 
 require("gpii-express");
-require("../../");
-
-require("gpii-test-browser");
-gpii.test.browser.loadTestingSupport();
+require("../../../");
 
 fluid.defaults("gpii.locationBar.tests.harness", {
     gradeNames: ["gpii.express"],
@@ -47,27 +48,6 @@ fluid.defaults("gpii.locationBar.tests.harness", {
                 path:    "/tests",
                 content: ["%gpii-location-bar-relay/tests"]
             }
-        }
-    }
-});
-
-// Our test caseHolder (based on a standard one from gpii-test-browser).
-fluid.defaults("gpii.locationBar.tests.caseHolder", {
-    gradeNames: ["gpii.test.browser.caseHolder.withExpress"]
-});
-
-// Our test environment (based on a standard one from gpii-test-browser).
-fluid.defaults("gpii.locationBar.tests.environment", {
-    gradeNames: ["gpii.test.browser.environment.withExpress"],
-    startUrl: {
-        expander: {
-            funcName: "fluid.stringTemplate",
-            args: ["%baseUrl%endpoint", { baseUrl: "{testEnvironment}.options.url", endpoint: "{testEnvironment}.options.endpoint"}]
-        }
-    },
-    components: {
-        express: {
-            type: "gpii.locationBar.tests.harness"
         }
     }
 });
