+++ tests/static/tests-locationBar-binder.html
@@ -16,6 +16,9 @@
 
         <script type="text/javascript" src ="/nm/gpii-binder/src/js/binder.js"></script>
 
+        <!-- gpii-express dependency -->
+        <script type="text/javascript" src="/nm/gpii-express/src/js/lib/querystring-coding.js"></script>
+
         <script type="text/javascript" src ="/src/js/locationBar.js"></script>
         <script type="text/javascript" src="/tests/static/js/tests-locationBar-binder.js"></script>
     </head>
