+++ tests/static/tests-locationBar.html
@@ -11,6 +11,9 @@
         <script type="text/javascript" src="/nm/infusion/src/framework/core/js/ModelTransformation.js"></script>
         <script type="text/javascript" src="/nm/infusion/src/framework/core/js/ModelTransformationTransforms.js"></script>
 
+        <!-- gpii-express dependency -->
+        <script type="text/javascript" src="/nm/gpii-express/src/js/lib/querystring-coding.js"></script>
+
         <script type="text/javascript" src ="/src/js/locationBar.js"></script>
         <script type="text/javascript" src="/tests/static/js/tests-locationBar.js"></script>
     </head>
